+++ src/modules/artworks/views/index.tsx
@@ -1,5 +1,5 @@
 import { fetchArtworks } from '~/modules/artworks/actions/fetchArtworks'
-import { ArtworksList } from '~/modules/artworks/views/components/ArtworkList'
+import { ArtworksList } from '~/modules/artworks/views/components/ArtworksList'
 
 export async function ArtworksView() {
   const initialData = await fetchArtworks({ page: 1 })
